Allow dotted field names in column hider
This allows dotted field names to be used in the column hider.  Since the table columns are hidden by applying a class based on the field name, they must be selected by using an attribute selector e.g. `$('[class="a.b.c"]')` instead of `$('.a.b.c`)` which gets misinterpreted for obvious reasons.  The other approach would be to escape/replace the dots when generating the class names, but this seemed easier.