Handle a non-existent field in an array
This was giving me an error, since I had several records for which a particular field was not defined.  Since I use a schema-less DB (MongoDB), this field doesn't even exist with a null value.  `data_get` performs an `array_key_exists` check before attempting to use `$part` as an index for the `$res` array.

This entire method could actually be replaced with data_get for simplicity, since this is what Laravel uses internally to retrieve values in dot notation.
```php
protected function extractCellValue($fieldName)
{
    try {
        return data_get($this->src, $fieldName);
    } catch (Exception $e) {
        throw new RuntimeException(
            "Can't read '$fieldName' property from DataRow",
            0,
            $e
        );
    }
}
```